refactor: move gocannon exported members to lib package
