+++ main.go
@@ -2,8 +2,16 @@ package main
 
 import (
 	"fmt"
+	"os"
 	"runtime"
+
+	"github.com/kffl/gocannon/lib"
 )
+
+func exitWithError(err error) {
+	fmt.Fprintf(os.Stderr, "error: %v\n", err)
+	os.Exit(1)
+}
 
 func main() {
 	config, err := parseArgs()
@@ -17,7 +25,7 @@ func main() {
 		printHeader(config)
 	}
 
-	g, err := NewGocannon(config)
+	g, err := lib.NewGocannon(config)
 
 	if err != nil {
 		exitWithError(err)
